requirements.txt
nonebot2
loguru
pydantic